modify observer pattern, integrate the event to abstract factory

+++ backend/src/main/java/com/sda_project/myfluffy/pet/dto/PetCreateDto.java
@@ -8,7 +8,7 @@ public class PetCreateDto {
 
     private String name;
 
-    private float age;
+    private double age;
 
     private String animalType;
