+++ backend/src/main/java/com/sda_project/myfluffy/pet/dto/PetDto.java
@@ -12,7 +12,7 @@ public class PetDto {
 
     private String name;
 
-    private float age;
+    private double age;
 
     private String animalType;
